Relax aws-sdk dependency versions (#54)
Specifying a fixed `0.10` only allows patch updates, which causes issues if trying to use for example `0.12` in a parent project, as handling of major version zero is different to one and up due to semver guarantees. ie. `0.10` is equivalent to `>=0.10.0, <0.11.0` where as `1.10` is equivalent to `>=1.10.0, <2.0.0`.

Also fix a couple minor clippy lints while I'm here.

Signed-off-by: Thomas Farr <xtansia@xtansia.com>